Test data are moved to separate .py files.
