+++ test/test_add_contact.py
@@ -1,18 +1,7 @@
 # -*- coding: utf-8 -*-
-from model.contact import Contact
 import pytest
-import random
-import string
-
-
-def rand_string(prefix, maxlen):
-    chars = string.ascii_letters + string.digits + string.punctuation + " " * 10
-    return prefix + "".join([random.choice(chars) for i in range(random.randrange(maxlen))])
-
-test_data = [Contact()] + [Contact(fname=rand_string("fname", 10), lname=rand_string("lname", 10),
-                                   address1=rand_string("address", 30), hphone=rand_string("homeph", 10),
-                                   mphone=rand_string("mobileph", 10), email1=rand_string("email1", 15))
-                           for i in range(2)]
+from model.contact import Contact
+from data.contact import const as test_data
 
 
 @pytest.mark.parametrize("contact", test_data, ids=[repr(x) for x in test_data])
