+++ test/test_add_group.py
@@ -1,18 +1,7 @@
 # -*- coding: utf-8 -*-
-from model.group import Group
 import pytest
-import random
-import string
-
-
-def rand_string(prefix, maxlen):
-    chars = string.ascii_letters + string.digits + string.punctuation + " " * 10
-    return prefix + "".join([random.choice(chars) for i in range(random.randrange(maxlen))])
-
-test_data = [Group(name=name, header=header, footer=footer)
-             for name in ("", rand_string("name", 10))
-             for header in ("", rand_string("header", 20))
-             for footer in ("", rand_string("footer", 20))]
+from model.group import Group
+from data.group import const as test_data
 
 
 @pytest.mark.parametrize("group", test_data, ids=[repr(x) for x in test_data])
